* origin/pr/14573:
  random sprites for player and NPCs

* origin/pr/14600:
  Suppress GCC warning
  Make recursion more concise/explicit
  Ensure we visit parent item only once
  Visit the item first followed by any contained items
  Revert "Visit the item first followed by any contained items"
  Visit the item first followed by any contained items
  Move VisitResponse enum to item.h and capitalise members
  Reimplement Character::has_item_with using visitor
  Reimplement player::has_items_with_quality using visitor
  Add const versions of visitor methods
  Implement visitor pattern for character possessions
  Implement visitor pattern for character inventory
  Implement visitor pattern for contained items

* origin/pr/14636:
  Provide missing difficulty arguments
  Fix training rate for dodging skill
  Use static variable
  Force dodge training to be positive rather than zero
  Simplify check for adjacent attackers in on_dodge()
  Require caller of on_dodge() to specify difficulty
  Remove unnecessary calls to on_dodge()
  Remove virtual function Creature::dodge_hit
  Move logic from player::dodge_hit to player::on_dodge

+++ src/weather.h
@@ -69,7 +69,7 @@ struct weather_animation_t {
 /**
  * Weather animation settings for the given type.
  */
-weather_animation_t get_weather_animation(weather_type type);
+weather_animation_t get_weather_animation( weather_type type );
 
 /**
  * Weather drawing tracking.
@@ -88,18 +88,19 @@ struct weather_printable {
  * Environmental effects and ramifications of weather.
  * Visibility range changes are done elsewhere.
  */
-namespace weather_effect {
-void none       (); //!< Fallback weather.
-void glare      ();
-void wet        ();
-void very_wet   ();
-void thunder    ();
-void lightning  ();
-void light_acid ();
-void acid       ();
-void flurry     (); //!< Currently flurries have no additional effects.
-void snow       (); //!< Currently snow has no additional effects.
-void snowstorm  (); //!< Currently snowstorms have no additional effects.
+namespace weather_effect
+{
+void none();        //!< Fallback weather.
+void glare();
+void wet();
+void very_wet();
+void thunder();
+void lightning();
+void light_acid();
+void acid();
+void flurry();      //!< Currently flurries have no additional effects.
+void snow();        //!< Currently snow has no additional effects.
+void snowstorm();   //!< Currently snowstorms have no additional effects.
 } //namespace weather_effect
 
 struct weather_datum {
@@ -110,7 +111,7 @@ struct weather_datum {
     int light_modifier;     //!< Modification to ambient light.
     int sound_attn;         //!< Sound attenuation of a given weather type.
     bool dangerous;         //!< If true, our activity gets interrupted.
-    void (*effect)();       //!< Function pointer for weather effects.
+    void ( *effect )();     //!< Function pointer for weather effects.
 };
 
 struct weather_sum {
@@ -119,9 +120,9 @@ struct weather_sum {
     float sunlight = 0.0f;
 };
 
-std::string const& season_name(int season);
-std::string const& season_name_upper(int season);
-weather_datum const& weather_data(weather_type type);
+std::string const &season_name( int season );
+std::string const &season_name_upper( int season );
+weather_datum const &weather_data( weather_type type );
 
 std::string weather_forecast( point const &abs_sm_pos );
 
@@ -131,14 +132,14 @@ std::string weather_forecast( point const &abs_sm_pos );
 // If scale is Fahrenheit: temperature(100) will return "100F"
 //
 // Use the decimals parameter to set number of decimal places returned in string.
-std::string print_temperature(float fahrenheit, int decimals = 0);
-std::string print_humidity(float humidity, int decimals = 0);
-std::string print_pressure(float pressure, int decimals = 0);
+std::string print_temperature( double fahrenheit, int decimals = 0 );
+std::string print_humidity( double humidity, int decimals = 0 );
+std::string print_pressure( double pressure, int decimals = 0 );
 
-int get_local_windchill(double temperature, double humidity, double windpower);
-int get_local_humidity(double humidity, weather_type weather, bool sheltered = false);
-int get_local_windpower(double windpower, std::string const &omtername = "no name",
-                        bool sheltered = false);
+int get_local_windchill( double temperature, double humidity, double windpower );
+int get_local_humidity( double humidity, weather_type weather, bool sheltered = false );
+int get_local_windpower( double windpower, std::string const &omtername = "no name",
+                         bool sheltered = false );
 
 weather_sum sum_conditions( const calendar &startturn,
                             const calendar &endturn,
@@ -150,7 +151,8 @@ weather_sum sum_conditions( const calendar &startturn,
  * by the @ref map, but absolute).
  * @param tr The funnel (trap which acts as a funnel).
  */
-void retroactively_fill_from_funnel( item &it, const trap &tr, const calendar &endturn, const tripoint &pos);
+void retroactively_fill_from_funnel( item &it, const trap &tr, const calendar &endturn,
+                                     const tripoint &pos );
 
 double funnel_charges_per_turn( double surface_area_mm2, double rain_depth_mm_per_hour );
